update:added authentication wrapper
