+++ koduyorum-web/src/App.js
@@ -10,6 +10,7 @@ import SearchResults from './SearchResults';
 import CodeExecution from './components/code-execution';
 import SurveyPage from './SurveyPage';
 import PostQuestion from './PostQuestion';
+import AuthWrapper from './AuthWrapper';
 
 
 const App = () => {
@@ -20,11 +21,11 @@ const App = () => {
         <Route path="/signup" element={<Signup />} />
         <Route path="/login" element={<Login />} />
         <Route path="/forgot" element={<ForgotPassword />} />
-        <Route path="/feed" element={<Feed />} />
-        <Route path="/question" element={<CodeExecution />} />
-        <Route path="/result/:wiki_id/:wiki_name" element={<SearchResults />} />
-        <Route path="/survey" element={<SurveyPage />} />
-        <Route path="/post_question" element={<PostQuestion />} />
+        <Route path="/feed" element={<AuthWrapper><Feed /></AuthWrapper>} />
+        <Route path="/question" element={<AuthWrapper><CodeExecution /></AuthWrapper>} />
+        <Route path="/result/:wiki_id/:wiki_name" element={<AuthWrapper><SearchResults /></AuthWrapper>} />
+        <Route path="/survey" element={<AuthWrapper><SurveyPage /></AuthWrapper>} />
+        <Route path="/post_question" element={<AuthWrapper><PostQuestion /></AuthWrapper>} />
       </Routes>
     </Router>
   );
